arm: msm: watchdogv2: export pet_watchdog interface
Falcon is using msm_watchdog_v2 to replace the original msm_watchdog,
but the msm_watchdog.h was not updated accordingly.

So that the pet_watchdog and msm_watchdog_reset interfaces are defined as null,
any callers to touch_hw_watchdog/pet_watchdog and msm_watchdog_reset would not
take effect.

The current callers include:
    apanic_mmc
    soft lockup detector
    modem panic/shutdown handler
    CPU power off handler

That is the reason that we see a lot watchdog reset happens during softlockup
and kernel panic dump.

Change-Id: Ib38c4d988fbd6b32d1aa24691b609f9f3eb812f2